Fixed issue with test names
